+++ lib/CMakeLists.txt
@@ -20,7 +20,6 @@ add_subdirectory(Basic)
 add_subdirectory(ClangImporter)
 add_subdirectory(Demangling)
 add_subdirectory(Driver)
-add_subdirectory(DWARFImporter)
 add_subdirectory(Frontend)
 add_subdirectory(FrontendTool)
 add_subdirectory(Index)
