+++ lib/ClangImporter/CMakeLists.txt
@@ -9,6 +9,7 @@ add_swift_host_library(swiftClangImporter STATIC
   ClangAdapter.cpp
   ClangDiagnosticConsumer.cpp
   ClangImporter.cpp
+  DWARFImporter.cpp
   IAMInference.cpp
   ImportDecl.cpp
   ImportEnumInfo.cpp
