--- lib/DWARFImporter/CMakeLists.txt
@@ -1,6 +1,0 @@
-add_swift_host_library(swiftDWARFImporter STATIC
-  DWARFImporter.cpp)
-target_link_libraries(swiftDWARFImporter PRIVATE
-    swiftAST
-    swiftParse
-)
