+++ lib/Frontend/Frontend.cpp
@@ -23,7 +23,6 @@
 #include "swift/Basic/FileTypes.h"
 #include "swift/Basic/SourceManager.h"
 #include "swift/Basic/Statistic.h"
-#include "swift/DWARFImporter/DWARFImporter.h"
 #include "swift/Frontend/ParseableInterfaceModuleLoader.h"
 #include "swift/Parse/DelayedParsingCallbacks.h"
 #include "swift/Parse/Lexer.h"
@@ -379,17 +378,6 @@ bool CompilerInstance::setUpModuleLoaders() {
   }
   Context->addModuleLoader(std::move(SML));
   Context->addModuleLoader(std::move(clangImporter), /*isClang*/ true);
-
-  if (Invocation.getLangOptions().EnableDWARFImporter) {
-    auto dwarfImporter =
-        DWARFImporter::create(*Context, Invocation.getClangImporterOptions(),
-                              nullptr, getDependencyTracker());
-    if (!dwarfImporter) {
-      Diagnostics.diagnose(SourceLoc(), diag::error_clang_importer_create_fail);
-      return true;
-    }
-    Context->addModuleLoader(std::move(dwarfImporter));
-  }
 
   return false;
 }
