--- test/DWARFImporter/Inputs/objc-header.h
@@ -1,7 +1,0 @@
-struct Point {
-  int x, y;
-};
-
-@interface ObjCClass
-- (instancetype)init;
-@end
