+++ tools/SourceKit/lib/SwiftLang/CMakeLists.txt
@@ -12,7 +12,7 @@ add_sourcekit_library(SourceKitSwiftLang
   SwiftTypeContextInfo.cpp
   LINK_LIBS
     SourceKitCore swiftDriver swiftFrontend
-    swiftDWARFImporter swiftClangImporter swiftIDE
+    swiftClangImporter swiftIDE
     swiftAST swiftMarkup swiftParse swiftParseSIL swiftSIL swiftSILGen
     swiftSILOptimizer swiftIRGen swiftSema swiftBasic swiftSerialization
     swiftSyntax swiftOption libcmark_static
