+++ tools/sil-func-extractor/CMakeLists.txt
@@ -5,7 +5,6 @@ add_swift_host_tool(sil-func-extractor
 target_link_libraries(sil-func-extractor
                       PRIVATE
                         swiftClangImporter
-                        swiftDWARFImporter
                         swiftFrontend
                         swiftSerialization
                         swiftSILGen
