+++ tools/sil-nm/CMakeLists.txt
@@ -5,7 +5,6 @@ add_swift_host_tool(sil-nm
 target_link_libraries(sil-nm
                       PRIVATE
                         swiftClangImporter
-                        swiftDWARFImporter
                         swiftFrontend
                         swiftSerialization
                         swiftSILGen
