+++ unittests/Parse/CMakeLists.txt
@@ -10,10 +10,8 @@ target_link_libraries(SwiftParseTests
     PRIVATE
     swiftSIL
     swiftClangImporter
-    swiftDWARFImporter
     swiftParse
     swiftAST
     # FIXME: Sema must go last because of circular dependencies with AST.
     swiftSema
 )
-
